add permissions for information_schema checks
added SELECT on *.* so information_schema can be checked and table sizes
reported.
note that *.*:SELECT is required since it seems that permissions on
information_schema are handled in a specific way, i.e. you can not check
values related to database X is you do not have access to X (at least in
information_schema.schemata).